Allow to specify lists of possible values as well.
One should be careful with specifying IdGroup - the function now
assumes that if a list is given, then it is a list of possible
values. So to specify that the value should be [8,3], one should
supply [[8,3]] as an argument. Otherwise, it will check if the
IdGroup is 8 or 3 and will return an empty list.